analysis-resume/pyproject.toml
[project]
name = "store-resume"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.38.0",
    "celery>=5.6.3",
    "langchain-openai>=1.1.0",
    "langchain-text-splitters>=0.3.0",
    "openai>=2.30.0",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.0",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
    "sqlalchemy>=2.0.0",
    "structlog>=25.5.0",
]
